+++ plugins/CircuitExplorer/io/bbploader/GIDLoader.cpp
@@ -81,9 +81,9 @@ class GIDSampler
 public:
     static brain::GIDSet subSample(const brain::GIDSet &input, float percentage)
     {
-        auto nbGids = input.size();
+        auto nbGids = static_cast<float>(input.size());
         const auto expectedSize = static_cast<size_t>(nbGids * percentage);
-        auto skipFactor = static_cast<size_t>(static_cast<float>(nbGids) / static_cast<float>(expectedSize));
+        auto skipFactor = static_cast<size_t>(nbGids / static_cast<float>(expectedSize));
         skipFactor = std::max(skipFactor, 1ul);
         brain::GIDSet subsampled;
         auto it = subsampled.begin();
